--- workspace/marqua/src/index.ts
@@ -1,2 +1,0 @@
-export { compile, parse, traverse } from './internal/index.js';
-export { forge } from './internal/config.js';
